Simplify `offscreen` API as `color` module in `iced_wgpu`
