+++ wgpu/src/lib.rs
@@ -46,7 +46,7 @@ pub mod geometry;
 
 mod backend;
 mod buffer;
-mod offscreen;
+mod color;
 mod quad;
 mod text;
 mod triangle;
